Altering homepage map visualization to utilize static json data - Updated homepage json file with the values to be used - Removed API call and subbed in the static data - Converted container into a component, as it no longer needs to be a container
