+++ src/graphics/homepage.json
@@ -67,5 +67,62 @@
         "#87dc85",
         "#95e784",
         "#a3f383"
+    ],
+"states": {
+    "children": [
+        {"item": "AL", "aggregate": 66910372684},
+        {"item": "AK", "aggregate": 7986325057},
+        {"item": "AZ", "aggregate": 71558643601},
+        {"item": "AR", "aggregate": 22925498923},
+        {"item": "AS", "aggregate": 266100668},
+        {"item": "CA", "aggregate": 289679921444},
+        {"item": "CO", "aggregate": 39484246622},
+        {"item": "CT", "aggregate": 74611923115},
+        {"item": "DE", "aggregate": 6627398656},
+        {"item": "DC", "aggregate": 30931763562},
+        {"item": "FL", "aggregate": 171054856475},
+        {"item": "GA", "aggregate": 65183957424},
+        {"item": "HI", "aggregate": 11118299699},
+        {"item": "ID", "aggregate": 12337454227},
+        {"item": "IL", "aggregate": 80558323346},
+        {"item": "IN", "aggregate": 128419186641},
+        {"item": "IA", "aggregate": 23217582228},
+        {"item": "KS", "aggregate": 16229795917},
+        {"item": "KY", "aggregate": 89303164343},
+        {"item": "LA", "aggregate": 35424256433},
+        {"item": "ME", "aggregate": 11660520983},
+        {"item": "MD", "aggregate": 66352632820},
+        {"item": "MA", "aggregate": 62367260228},
+        {"item": "MI", "aggregate": 77396050297},
+        {"item": "MN", "aggregate": 81535424896},
+        {"item": "MS", "aggregate": 24481678151},
+        {"item": "MO", "aggregate": 52971889116},
+        {"item": "MT", "aggregate": 8333041078},
+        {"item": "NE", "aggregate": 12306044107},
+        {"item": "NV", "aggregate": 17696587833},
+        {"item": "NH", "aggregate": 9706435188},
+        {"item": "NJ", "aggregate": 60811390189},
+        {"item": "NM", "aggregate": 22830683804},
+        {"item": "NY", "aggregate": 158617348110},
+        {"item": "NC", "aggregate": 63968795626},
+        {"item": "ND", "aggregate": 62603330175},
+        {"item": "OH", "aggregate": 76119620127},
+        {"item": "OK", "aggregate": 28215223114},
+        {"item": "OR", "aggregate": 32042458571},
+        {"item": "PA", "aggregate": 193604489691},
+        {"item": "RI", "aggregate": 8973218138},
+        {"item": "SC", "aggregate": 77700699550},
+        {"item": "SD", "aggregate": 6763485958},
+        {"item": "TN", "aggregate": 73841483061},
+        {"item": "TX", "aggregate": 169415582466},
+        {"item": "UT", "aggregate": 14543595370},
+        {"item": "VT", "aggregate": 5485620261},
+        {"item": "VA", "aggregate": 94665994579},
+        {"item": "WA", "aggregate": 58152788598},
+        {"item": "WV", "aggregate": 15265354160},
+        {"item": "WI", "aggregate": 84175926266},
+        {"item": "WY", "aggregate": 4146856654},
+        {"item": "GU", "aggregate": 1220551421}
     ]
+    }
 }
